disabled amp again

+++ op/fused_act.py
@@ -47,7 +47,8 @@ class FusedLeakyReLUFunction(Function):
     def forward(ctx, input, bias, negative_slope, scale):
         empty = input.new_empty(0)
         # print(input.dtype, bias.dtype, empty.dtype, 3, 0, negative_slope, scale)
-        out = fused.fused_bias_act(input.half(), bias.half(), empty.half(), 3, 0, negative_slope, scale)
+        # out = fused.fused_bias_act(input.half(), bias.half(), empty.half(), 3, 0, negative_slope, scale)
+        out = fused.fused_bias_act(input, bias, empty, 3, 0, negative_slope, scale)
         ctx.save_for_backward(out)
         ctx.negative_slope = negative_slope
         ctx.scale = scale
